Unify AI services onto one shared client (lib/ai-client.ts)
Eight AI services each redefined the same Firebase config + getAIModel (and
grounded-model) factory. Extract one shared lib/ai-client.ts (getAIModel,
getGroundedAIModel) and route all of them through it — removes ~7 duplicate
configs and 11 duplicate factory functions. No behavior change.
