+++ app/apps/game-analytics/lib/ai-queue-service.ts
@@ -1,23 +1,7 @@
 'use client';
 
-import { getAI, getGenerativeModel, GoogleAIBackend } from 'firebase/ai';
-import { initializeApp, getApps } from 'firebase/app';
+import { getAIModel } from './ai-client';
 import { QueueAIContext } from './calculations';
-
-const firebaseConfig = {
-  apiKey: "AIzaSyBS3IVvszDrm_zjjXu8TATgs1H-FlegHtM",
-  authDomain: "oneapp-943e3.firebaseapp.com",
-  projectId: "oneapp-943e3",
-  storageBucket: "oneapp-943e3.firebasestorage.app",
-  messagingSenderId: "1052736128978",
-  appId: "1:1052736128978:web:9d42b47c6a343eac35aa0b",
-};
-
-function getAIModel() {
-  const app = getApps().length ? getApps()[0] : initializeApp(firebaseConfig);
-  const ai = getAI(app, { backend: new GoogleAIBackend() });
-  return getGenerativeModel(ai, { model: "gemini-2.5-flash" });
-}
 
 // Cache key for localStorage
 const CACHE_KEY = 'queue-ai-cache';
