+++ app/apps/game-analytics/lib/subscription-games-service.ts
@@ -1,29 +1,8 @@
 'use client';
 
-import { getAI, getGenerativeModel, GoogleAIBackend } from 'firebase/ai';
-import { initializeApp, getApps } from 'firebase/app';
+import { getGroundedAIModel as getGroundedModel } from './ai-client';
 import { SubscriptionTier, SubscriptionBucket } from './types';
 import { monthLabel } from './subscription-settings';
-
-const firebaseConfig = {
-  apiKey: "AIzaSyBS3IVvszDrm_zjjXu8TATgs1H-FlegHtM",
-  authDomain: "oneapp-943e3.firebaseapp.com",
-  projectId: "oneapp-943e3",
-  storageBucket: "oneapp-943e3.firebasestorage.app",
-  messagingSenderId: "1052736128978",
-  appId: "1:1052736128978:web:9d42b47c6a343eac35aa0b",
-};
-
-// Grounded model — uses Google Search so it can find the real PS Plus lineup
-// for a given month and cite the post it pulled it from.
-function getGroundedModel() {
-  const app = getApps().length ? getApps()[0] : initializeApp(firebaseConfig);
-  const ai = getAI(app, { backend: new GoogleAIBackend() });
-  return getGenerativeModel(ai, {
-    model: 'gemini-2.5-flash',
-    tools: [{ googleSearch: {} }],
-  } as Parameters<typeof getGenerativeModel>[1]);
-}
 
 export interface SubscriptionGameItem {
   name: string;
